% Generated by roxygen2 (4.0.1): do not edit by hand
\name{anova_test}
\alias{anova_test}
\title{this function calculates the p val of the genotype variable according to anova
of two glmer fits or two glm fits}
\usage{
anova_test(data, bin, test, silent = T, log_file, tail, null.model = NULL,
  fit.null = FALSE)
}
\description{
this function calculates the p val of the genotype variable according to anova
of two glmer fits or two glm fits
}
\keyword{internal}

