% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_tail}
\alias{convert_tail}
\title{this function converts a two-tailed p value to a one-tailed p value}
\usage{
convert_tail(direction, pval, tail)
}
\arguments{
\item{direction}{is positive}

\item{pval}{pvalue}

\item{tail}{"decreasing", "increasing" or "both"}
}
\description{
this function converts a two-tailed p value to a one-tailed p value
}

