% Generated by roxygen2 (4.0.1): do not edit by hand
\name{double.link}
\alias{double.link}
\title{Statistical analysis for testing the combinatory effect of two mutations}
\usage{
double.link(vcfFile, pedFile, pheno.name, output = ".", test = "woG2",
  detect = "never", silent = TRUE, tail = "decreasing", prefix = "",
  cutoff_single = 0.01, plot.it = TRUE, transform.pheno = NULL,
  log.level = "WARN")
}
\arguments{
\item{vcfFile}{genotype input file in VCF format}

\item{pedFile}{pedigree file in PED format}

\item{pheno.name}{phenotype name to analyze}

\item{output}{The output folder to put all the output files of the analysis}

\item{test}{The statistical test to be used for identifying significant genes
associated with phenotype. Default is "woG2", without considering G2 mother
effect. Another option is "wG2", considering that effect. "wG2" will take a
long time.}

\item{detect}{A character string specifying whether to detect clustering of
phenotypic scores and to transform continuous phenotypic scores into a binary
variable (affected and nonaffected). This parameter only works on continuous
phenotype scores. "never" (default): never transform; "always": always
transform; "auto": let the program decide whether to transform.}

\item{silent}{Print intermediate messages to stdout if set to TRUE.}

\item{tail}{Either "decreasing", "increasing" or "both". "decreasing" tests
whether the mutation leads to decreased antibody reaction (default);
"increasing" tests whether the mutation leads to increased antibody reaction;
"both" tests the deviation in either direction.}

\item{cutoff_single}{A numerical value smaller than 1. When between 0-1, only
genes whose single effect has a p value larger than \code{cutoff_single} will
be considered in testing for combinatory effect. Setting \code{cutoff_single}
to a negative number will turn off this screening (not recommended). Default:
0.01.}

\item{prefix}{Default is "". An optional character string to be attached to
output file names. This can create personalized names for each job.}

\item{plot.it}{Default is TRUE, it controls whether to output linkage plots
and distribution plots.}

\item{transform.pheno}{Default is NULL. Use "log" if phenotypes need to log
transformed.}

\item{log.level}{Default WARN, but can be set from 'DEBUG', 'INFO', 'WARN'
and 'ERROR'.}
}
\value{
a list of two values will be returned. One is returncode (0: success).
The other is analysis results.
}
\description{
This function tests whether the combination of two mutations
could be correlated with phenotype.
}
\details{
The \code{double.link} function has 4 modes:
recessive, additive, dominant and inhibitory. Refer to the tutorial attached
in the package for the assumption in each mode. Recessive, additive, dominant
and inhibitory modes test whether the interaction between two genes could
explain the phenotypic variation with different assumptions of how the
interaction will take place.  Basically, ANOVA tests are applied with
different assumption in each mode to test the combinatory effects of two
genes. \code{cutoff_single} is suggested to be set to a value between 0-1, in
order to mask genes which themselves are already shown to be significantly
correlated with phenotype by \code{single.link}.  A PDF file
\code{linkage_plot} will be generated. It will contain 4 heatmap plots of p
values in each of the four modes. The p values are transformed to a -log10
scale. 4 CSV files containing the matrix of p values for each of the 4 modes
will be generated. A second PDF file \code{distribution_plot} will be
generated containing scattorplots of phenotypes scores or categorical
phenotypes for each pair of genes deemed to be significant. A TXT file
\code{log} containing messages, warnings and errors of the program will be
generated.  Synthetic lethality is tested in order to find whether G3 mice
with genotype of VAR,VAR; VAR,HET; and HET,VAR will have any decreased chance
of survival. The p values for synthetic lethality are presented along with
the p values for combinatory effect.
}
\examples{
path <- system.file("extdata/double",package="LinkageAnalysis")
vcfFile <- file.path(path, "R0491_body_weight.vcf")
pedFile <- file.path(path, "R0491_body_weight.ped")
pheno.name <- "weight"
output <- file.path(path, "output")
ret <- double.link(vcfFile, pedFile, pheno.name, output,
                   test = "woG2", tail = "both", prefix="double")
}
\seealso{
\code{\link{single.link}}.
}

