% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.vcf}
\alias{get.vcf}
\title{Read a VCF file}
\usage{
get.vcf(vcfFile, log_file)
}
\arguments{
\item{vcfFile}{file name in VCF format}

\item{log_file}{specify a log file}
}
\value{
a list of parsed VCF file content
}
\description{
Read a VCF file
}

