% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_data}
\alias{get_data}
\title{this function reads the main data and the optional G2 genotype data if G2
genotype data, the returned list will have an additional element}
\usage{
get_data(main_file = "", G2 = "", log_file, detect,
  transform.pheno = NULL)
}
\arguments{
\item{main_file}{a CSV file listing phenotyping and genotypes of G3 mice}

\item{G2}{a CSV file listing G2 mothers' genotypes}

\item{log_file}{specify log file}

\item{detect}{"auto" means detect binary phenotype, or "NULL" means do nothing}

\item{transform.pheno}{"log" apply log transformation, or NULL means do nothing}
}
\value{
a list containing:
  returncode: 0 when success
  message: optionally message
  data
    genes: data.frame (sorted by chromsomal positions)
      Gene: gene names, character vector
      Coordination: character vector, e.g. "1_123"
      chr: character vector
      pos: numeric vector
    phenotype: a list
      mother: factor vector, mother names
      sex: 0 = female, 1 = male, 0.5 = unknown
      phenotype: numeric vector
    genotype: character matrix with rowname = gene, colname = sampleName, e.g. "REF"
    bin: logical
    unconverted: list (may be NULL)
      phenotype: original phenotype
      break_point: break point
    G2: character matrix. After fixing genotypes, it's similar to G2 file
      Gene: char vec
      Coordination: char vec
      Amplicon: char vec
    n:  # of genes
    obs:  # of mice
}
\description{
this function reads the main data and the optional G2 genotype data if G2
genotype data, the returned list will have an additional element
}
\keyword{internal}

