% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_main}
\alias{get_main}
\title{this function reads the main data}
\usage{
get_main(main_file, log_file, detect, transform.pheno = NULL)
}
\arguments{
\item{main_file}{a CSV file listing phenotyping and genotypes of G3 mice}

\item{log_file}{specify log file}

\item{detect}{"auto" means detect binary phenotype, or "NULL" means do nothing}

\item{transform.pheno}{"log" apply log transformation, or NULL means do nothing}
}
\value{
NULL when criitical error
}
\description{
this function reads the main data
}

