% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{LinkageAnalysis}
\alias{LinkageAnalysis}
\alias{LinkageAnalysis-package}
\title{Linkage Analysis package}
\description{
This package collects statistical analysis procedures to
acess the significance of correlation with phenotype and homozygous
lethality for singel gene and a pair of genes.
}
\references{
https://mutagenetix.utsouthwestern.edu
}

