% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mht}
\alias{mht}
\title{this function draws the manhattan plot}
\usage{
mht(genes, sig_gene, type, log_file, test, bin, effective, genotype = NULL)
}
\arguments{
\item{genes}{get_data()$genes, including Gene, Coordination, chr, pos, REF, HET, VAR, lethal}

\item{sig_gene}{num_gene by 4 columns ("additive", "recessive", "dominant", "TDT")}

\item{type}{any one of the "additive", "recessive", "dominant", "TDT"}

\item{log_file}{character}

\item{test}{"wG2" or "woG2"}

\item{bin}{logical}

\item{effective}{logical vector}

\item{genotype}{data.frame, as in get_data()$genotype}
}
\description{
this function draws the manhattan plot
}

