% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.manhattan}
\alias{plot.manhattan}
\title{Draw manhattan plot}
\usage{
\method{plot}{manhattan}(data, main = "")
}
\arguments{
\item{data}{a data frame with at least three columns: Chrom, Position, Pval}

\item{main}{plot title}
}
\description{
Draw manhattan plot
}

