% Generated by roxygen2 (4.0.1): do not edit by hand
\name{single.lethal.getPvalue}
\alias{single.lethal.getPvalue}
\title{Calculate single lethal p-value}
\usage{
single.lethal.getPvalue(nvarFromHet, nFromHet, nvarFromUnknown, nFromUnknown)
}
\arguments{
\item{nvarFromHet}{counts of G3 genotyped and with HET mom}

\item{nFromHet}{counts of G3 with VAR and with HET mom}

\item{nvarFromUnknown}{counts of offsprintgs genotyped and with Unknown mom}

\item{nFromUnknown}{counts of G3 with VAR and with Unknown mom}
}
\description{
Calculate single lethal p-value
}

