% Generated by roxygen2 (4.0.1): do not edit by hand
\name{single_link}
\alias{single_link}
\title{Statistical analysis for correlation with phenotype and lethality on a single locus}
\usage{
single_link(main_file = "", G2_file = "", output = ".", test = "wG2",
  detect = "never", silent = T, tail = "decreasing", prefix = "",
  plot.it = TRUE, transform.pheno = NULL)
}
\arguments{
\item{main_file}{\code{raw_file} is an csv file containing main data.
See \code{details} for more information.}

\item{G2_file}{\code{G2_file} is an optional csv file containing G2 genotype
data, if such data are available. See \code{details} for more information.}

\item{output}{The output folder to put all the output files of the analysis}

\item{test}{The statistical test to be used for identifying significant genes
associated with phenotype. Default is "wG2", considering G2 mother effect.
Another option is "woG2", not considering that effect.}

\item{detect}{A character string specifying whether to detect clustering of
phenotypic scores and to transform continuous phenotypic scores into a binary
variable (affected and nonaffected). This parameter only works on continuous
phenotype scores. "never" (default): never transform; "always": always
transform; "auto": let the program decide whether to transform.}

\item{silent}{Print intermediate messages to stdout if set to TRUE.}

\item{tail}{Either "decreasing", "increasing" or "both". "decreasing" tests
whether the mutation leads to decreased antibody reaction (default);
"increasing" tests whether the mutation leads to increased antibody reaction;
"both" tests the deviation in either direction.}

\item{prefix}{Default is "". An optional character string to be attached to
output file names. This can create personalized names for each job.}

\item{plot.it}{Default is TRUE, it controls whether to output linkage plots
and distribution plots.}

\item{transform.pheno}{Default is NULL. Use "log" if phenotypes need to log
transformed.}
}
\value{
No return value.
}
\description{
This function reads the raw data and conducts statistical
analysis for the significance of correlation with phenotype and homozygous
lethality for each gene.
}
\details{
First of all, this package depends on \code{lme4}. The current
version of this package on CRAN does not seem to be a stable one (at least on
my computer). I put an archived version of the lme4 package in the
\code{inst/extdata} folder of the \code{LinkageAnalysis} package. It works
well on my Linux system. You can install it from source on your Linux
platform, too. If you are working on PC/Mac, then you may also need to search
for one version of the \code{lme4} package that is stable on your machine.
\code{main_file} is a csv file with fixed format. The six cells in A1-B3 can
be anything. A4 must be "Gene", B4 must be "Coordination", C4 must be
"Amplicon", C3 must be "Phenotype", C2 must be "G3 gender" and C1 must be "G2
Dam eartag". The cells below A4 are gene names. The cells below B4 are
genomic locations ("1_100", for example, means chr1 and the 100th bp). The
cells below C4 are Amplicon. The cells to the right of C4 G3 eartag. The
cells to the right of C3 are phenotypes. They can be categorical values like,
"Affected", "Intermediate" and "Unaffected" or they can be numerical values,
but can not be a mixture of both. The cells to the right of C2 are G3 gender
and must be one of "F" and "M". The cells to the right of C1 are "G2 Dam
eartag". The cells in the bottom-right area are the genotypes. "REF", "HET"
and "VAR" correspond to wild-type, heterozygous mutant and homozygous
mutant. Other symbols will be treated as NA values. A sample raw input file
is stored in the \code{inst/extdata} folder of this package.  \code{G2_file}
is an optional csv file with fixed format. A1 must be "Gene", B1 must be
"Coordination" and C1 must be "Amplicon". The cells below A1 are gene
names. The cells below B1 are genomic locations. The cells below C1 are
Amplicon. The cells to the right of C1 are G2 dam names. The cells below G2
dam names are the G2 genotypes. "REF", "HET" and "VAR" correspond to
wild-type, heterozygous mutant and homozygous mutant. Other symbols will be
treated as NA values. A sample raw input file is stored in the
\code{inst/extdata} folder of this package. The \code{G2_file} can contain
more or less G2 dam mice and more or less genes than the
\code{main_file}. The algorithm will calculate the p values for lethality
according to different algorithms based on whether G2 dam mice are available
for a certain gene. The final p value is the combined p values from all the
G3 mice.  A PDF file will be generated. It will contain a diagnostic plot if
continuous phenotype scores are provided in the \code{main_file}. It will
contain another three manhattan plots, each assuming one of the three
"additive", "recessive" and "dominant" models. The two cutoff values are 0.05
and 0.05/(number of genes). The p values are transformed to a -log10
scale. It will also contain a plot for the analysis result of TDT analysis.
A CSV file containing will be generated. Columns names are in the first
line. The REF, HET, VAR columns are the number of mice with each genotype for
each gene. The "lethal" column is the p value for testing homozygous
lethality. The "additive", "recessive" and "dominant" columns are the p
values for testing the significance of the correlation of each gene and the
phenotype under different models. The "TDT" column is the p value of TDT test
for testing each gene's significance. The "Penetrance_REF", "Penetrance_HET",
"Penetrance_VAR" columns are the proportions of G3 mice with each genotype
showing the "affected" phenotype (only applies to binary phenotype scores or
binary scores that are converted from continuous phenotype scores). The
"Semidominance" column is the semidominance of the mutations, only calculated
when continuous phenotype scores are
provided. Semidominance=(HET-REF)/(VAR-REF). 0 means the HET phenotype is the
same as or over the average of the REF phenotype and 1 means the HET
phenotype is the same as or over the average of the VAR phenotype. Note that
VAR may not necessarily be lower than REF.  A second PDF file will be
generated containing scattorplots of phenotypes scores (continuous variable)
or table plots of phenotypes (binary variable).  A TXT file containing
messages, warnings and errors of the program will be generated.  The
statistical model used to calculate the p value of lethality, more accurately
speaking, is testing "whether homozygous mutation will have any detrimental
effect on the survival of mice at the time of data collection". chrX genes
are ignored for the moment.
}
\examples{
main_file <-
  system.file("extdata/Aquamarine_QM_Report.csv",package="LinkageAnalysis")
main_file_continuous <-
  system.file("extdata/Aquamarine_QM_Report_continuous.csv",package="LinkageAnalysis")
G2_file <-
  system.file("extdata/Aquamarine_QM_Report_G2.csv",package="LinkageAnalysis")
output <-
  sub(pattern="Aquamarine_QM_Report.csv",replacement="output",x=main_file)

# if categorical phenotype scores are given
single_link(main_file,G2_file,output=output,silent=TRUE,prefix="single")
# if continuous numerical phenotype scores are given
single_link(main_file_continuous,G2_file,output=output,silent=TRUE,prefix="single")
}

