% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vcf.delete.sample.by.index}
\alias{vcf.delete.sample.by.index}
\title{Delete samples to vcf data}
\usage{
vcf.delete.sample.by.index(vcf, index)
}
\arguments{
\item{vcf}{vcf data}

\item{index}{integer vector, delete those samples}
}
\value{
vcf data
}
\description{
Delete samples to vcf data
}

